add flake8 and black CI, reformat with black

+++ utils/helper_functions.py
@@ -2,6 +2,7 @@
 Seqlet extraction from MoDISco HDF5, coordinate utilities, and
 per-sequence motif-hit plotting (logo + gene map).
 """
+
 import os
 import logging
 import numpy as np
@@ -22,9 +23,10 @@ logging.basicConfig(level=logging.INFO)
 # Seqlet Extraction
 # ---------------------------------------------------------------------------
 
+
 def extract_seqlets_to_bed_modiscolite(
     h5_file: str,
-    pattern_group: str = 'both',
+    pattern_group: str = "both",
     example_names=None,
     indexing_df=None,
 ) -> pd.DataFrame:
@@ -43,19 +45,19 @@ def extract_seqlets_to_bed_modiscolite(
         Used to look up the native strand per sequence.
     """
     bed_rows = []
-    with h5py.File(h5_file, 'r') as f:
+    with h5py.File(h5_file, "r") as f:
         available = set(f.keys())
 
         def resolve_groups(pg):
-            if pg in (None, 'both', 'all', 'patterns'):
-                groups = [g for g in ('pos_patterns', 'neg_patterns') if g in available]
+            if pg in (None, "both", "all", "patterns"):
+                groups = [g for g in ("pos_patterns", "neg_patterns") if g in available]
                 if not groups:
-                    groups = [g for g in available if g.endswith('_patterns')]
+                    groups = [g for g in available if g.endswith("_patterns")]
                 return groups
-            if pg in ('pos', 'pos_patterns'):
-                return [g for g in ('pos_patterns',) if g in available]
-            if pg in ('neg', 'neg_patterns'):
-                return [g for g in ('neg_patterns',) if g in available]
+            if pg in ("pos", "pos_patterns"):
+                return [g for g in ("pos_patterns",) if g in available]
+            if pg in ("neg", "neg_patterns"):
+                return [g for g in ("neg_patterns",) if g in available]
             return [pg] if pg in available else []
 
         groups = resolve_groups(pattern_group)
@@ -67,36 +69,52 @@ def extract_seqlets_to_bed_modiscolite(
 
         for group_name in groups:
             group = f[group_name]
-            prefix = 'pos' if group_name.startswith('pos') else 'neg'
+            prefix = "pos" if group_name.startswith("pos") else "neg"
             for pattern_name in group.keys():
-                seqlets    = group[pattern_name]['seqlets']
-                starts     = seqlets['start'][:]
-                ends       = seqlets['end'][:]
-                strands    = seqlets['is_revcomp'][:]
-                example_idxs = seqlets['example_idx'][:]
+                seqlets = group[pattern_name]["seqlets"]
+                starts = seqlets["start"][:]
+                ends = seqlets["end"][:]
+                strands = seqlets["is_revcomp"][:]
+                example_idxs = seqlets["example_idx"][:]
                 for i in range(len(starts)):
                     example = int(example_idxs[i])
-                    chrom   = example_names[example] if example_names else example
+                    chrom = example_names[example] if example_names else example
                     if indexing_df is not None:
                         native_strand = indexing_df.loc[
-                            indexing_df['index'] == example, 'strand'
+                            indexing_df["index"] == example, "strand"
                         ].values[0]
                     else:
-                        native_strand = '+'
+                        native_strand = "+"
                     strand = (
-                        ('-' if native_strand == '+' else '+')
-                        if strands[i] else native_strand
+                        ("-" if native_strand == "+" else "+")
+                        if strands[i]
+                        else native_strand
                     )
-                    bed_rows.append([
-                        chrom, int(starts[i]), int(ends[i]),
-                        f"{pattern_name}_seqlet_{i}", 0, strand,
-                        f"{prefix}_{pattern_name}", bool(strands[i]),
-                    ])
+                    bed_rows.append(
+                        [
+                            chrom,
+                            int(starts[i]),
+                            int(ends[i]),
+                            f"{pattern_name}_seqlet_{i}",
+                            0,
+                            strand,
+                            f"{prefix}_{pattern_name}",
+                            bool(strands[i]),
+                        ]
+                    )
 
     return pd.DataFrame(
         bed_rows,
-        columns=['example_index', 'start', 'end', 'name', 'score',
-                 'strand', 'pattern_label', 'is_revcomp'],
+        columns=[
+            "example_index",
+            "start",
+            "end",
+            "name",
+            "score",
+            "strand",
+            "pattern_label",
+            "is_revcomp",
+        ],
     )
 
 
@@ -104,16 +122,18 @@ def extract_seqlets_to_bed_modiscolite(
 # Coordinate Utilities
 # ---------------------------------------------------------------------------
 
+
 def update_coordinates(df: pd.DataFrame, tensor_start: int) -> pd.DataFrame:
     """Shift seqlet start/end by tensor_start."""
     df = df.copy()
     df["start"] = df["start"] + tensor_start
-    df["end"]   = df["end"]   + tensor_start
+    df["end"] = df["end"] + tensor_start
     return df
 
 
-def filter_by_strand(df: pd.DataFrame, seq_name: str,
-                      indexing_df: pd.DataFrame) -> pd.DataFrame:
+def filter_by_strand(
+    df: pd.DataFrame, seq_name: str, indexing_df: pd.DataFrame
+) -> pd.DataFrame:
     """Keep only rows matching the native strand for seq_name."""
     correct_strand = indexing_df.loc[
         indexing_df["unique_ID"] == seq_name, "strand"
@@ -123,14 +143,23 @@ def filter_by_strand(df: pd.DataFrame, seq_name: str,
 
 def reorder_modiscolite_columns(df: pd.DataFrame) -> pd.DataFrame:
     """Standardise column order for MoDISco-lite output."""
-    cols = ['pattern_label', 'example_index', 'start', 'end',
-            'name', 'score', 'strand', 'unique_ID']
+    cols = [
+        "pattern_label",
+        "example_index",
+        "start",
+        "end",
+        "name",
+        "score",
+        "strand",
+        "unique_ID",
+    ]
     return df[cols]
 
 
 # ---------------------------------------------------------------------------
 # Motif Hit Plotting — FIMO
 # ---------------------------------------------------------------------------
+
 
 def handle_fimo(
     motifs_of_interest_dict: dict,
@@ -143,7 +172,9 @@ def handle_fimo(
     """
     Run FIMO on primary_seq and generate a logo+gene-map plot for each hit.
     """
-    annotations_list = fimo(motifs=motifs_of_interest_dict, sequences=primary_seq, dim=1)
+    annotations_list = fimo(
+        motifs=motifs_of_interest_dict, sequences=primary_seq, dim=1
+    )
     all_rows = []
     for df in annotations_list:
         if "sequence_name" in df.columns and not df.empty:
@@ -154,9 +185,9 @@ def handle_fimo(
             if not tensor_start.empty:
                 df = update_coordinates(df, tensor_start.values[0])
                 df = filter_by_strand(df, seq_name, indexing_df)
-                df['unique_ID'] = seq_name
+                df["unique_ID"] = seq_name
                 if "motif_name" not in df.columns:
-                    df['motif_name'] = "unknown"
+                    df["motif_name"] = "unknown"
                 all_rows.append(df)
 
     if not all_rows:
@@ -164,16 +195,16 @@ def handle_fimo(
         return
 
     annotations = pd.concat(all_rows, ignore_index=True)
-    for de_novo_motif in annotations['motif_name'].unique():
-        motif_df = annotations[annotations['motif_name'] == de_novo_motif]
+    for de_novo_motif in annotations["motif_name"].unique():
+        motif_df = annotations[annotations["motif_name"] == de_novo_motif]
         motif_logo_plot_dir = os.path.join(logo_plot_dir, de_novo_motif)
         os.makedirs(motif_logo_plot_dir, exist_ok=True)
-        for seq_name in motif_df['unique_ID'].unique():
-            df = motif_df[motif_df['unique_ID'] == seq_name]
+        for seq_name in motif_df["unique_ID"].unique():
+            df = motif_df[motif_df["unique_ID"] == seq_name]
             if not df.empty:
-                top_row  = df.loc[df['score'].idxmax()]
+                top_row = df.loc[df["score"].idxmax()]
                 attr_idx = indexing_df.loc[
-                    indexing_df["unique_ID"] == seq_name, 'index'
+                    indexing_df["unique_ID"] == seq_name, "index"
                 ].values[0]
                 arr = pt_attributions[attr_idx]
                 plot_logo_and_optional_gene_map(
@@ -184,6 +215,7 @@ def handle_fimo(
 # ---------------------------------------------------------------------------
 # Motif Hit Plotting — MoDISco-lite
 # ---------------------------------------------------------------------------
+
 
 def handle_modiscolite(
     h5_file: str,
@@ -196,7 +228,7 @@ def handle_modiscolite(
     MODISCO_WINDOW: int = 5000,
     intron_of_interest: str = None,
     bw_files: dict = None,
-    y_scale: str = 'log1p',
+    y_scale: str = "log1p",
     figsize: tuple = (20, 1.5),
     **kwargs,
 ):
@@ -236,11 +268,11 @@ def handle_modiscolite(
         kwargs.pop(k, None)
 
     annotations = extract_seqlets_to_bed_modiscolite(
-        h5_file, 'both', indexing_df=indexing_df
+        h5_file, "both", indexing_df=indexing_df
     )
-    index_to_name = dict(zip(indexing_df['index'], indexing_df['unique_ID']))
-    annotations['unique_ID'] = (
-        annotations['example_index'].map(index_to_name).fillna('unknown')
+    index_to_name = dict(zip(indexing_df["index"], indexing_df["unique_ID"]))
+    annotations["unique_ID"] = (
+        annotations["example_index"].map(index_to_name).fillna("unknown")
     )
 
     # Shift seqlet coords from modisco-window space to full input-tensor space
@@ -249,40 +281,46 @@ def handle_modiscolite(
 
     # Filter to valid sequences and correct strand
     annotations = annotations[
-        annotations.apply(lambda row: (
-            row['unique_ID'] in indexing_df['unique_ID'].values and
-            row['strand'] == indexing_df.loc[
-                indexing_df['unique_ID'] == row['unique_ID'], 'strand'
-            ].values[0]
-        ), axis=1)
+        annotations.apply(
+            lambda row: (
+                row["unique_ID"] in indexing_df["unique_ID"].values
+                and row["strand"]
+                == indexing_df.loc[
+                    indexing_df["unique_ID"] == row["unique_ID"], "strand"
+                ].values[0]
+            ),
+            axis=1,
+        )
     ]
 
-    annotations['score'] = annotations['score'].astype(float)
-    annotations = annotations[annotations['pattern_label'].isin(motifs_of_interest)]
+    annotations["score"] = annotations["score"].astype(float)
+    annotations = annotations[annotations["pattern_label"].isin(motifs_of_interest)]
     annotations = reorder_modiscolite_columns(annotations)
 
     if intron_of_interest:
-        annotations = annotations[annotations['unique_ID'] == intron_of_interest]
+        annotations = annotations[annotations["unique_ID"] == intron_of_interest]
 
-    for de_novo_motif in annotations['pattern_label'].unique():
-        motif_df = annotations[annotations['pattern_label'] == de_novo_motif]
+    for de_novo_motif in annotations["pattern_label"].unique():
+        motif_df = annotations[annotations["pattern_label"] == de_novo_motif]
         motif_df.to_csv(
             os.path.join(logo_plot_dir, f"{de_novo_motif}_hits.csv"), index=False
         )
         motif_logo_plot_dir = os.path.join(logo_plot_dir, de_novo_motif)
         os.makedirs(motif_logo_plot_dir, exist_ok=True)
 
-        for seq_name in motif_df['unique_ID'].unique():
-            df = motif_df[motif_df['unique_ID'] == seq_name]
+        for seq_name in motif_df["unique_ID"].unique():
+            df = motif_df[motif_df["unique_ID"] == seq_name]
             if df.empty:
                 continue
-            top_row  = df.loc[df['score'].idxmax()]
+            top_row = df.loc[df["score"].idxmax()]
             attr_idx = indexing_df.loc[
-                indexing_df["unique_ID"] == seq_name, 'index'
+                indexing_df["unique_ID"] == seq_name, "index"
             ].values[0]
             arr = pt_attributions[attr_idx]
-            logger.info(f"Plotting {seq_name} | pattern {de_novo_motif} | "
-                        f"attr_sum={float(np.sum(arr)):.4e}")
+            logger.info(
+                f"Plotting {seq_name} | pattern {de_novo_motif} | "
+                f"attr_sum={float(np.sum(arr)):.4e}"
+            )
             plot_logo_gene_map_and_read_densities(
                 arr=arr,
                 row=top_row,
